Fix Min(ute) counting bug when there is no preceding focal
In the rare case where a point line is recorded before ANY focals had
yet been recorded, the SQL writer was crashing because there was no
focal to associate the point with.  Fixed so that these orphan points
don't cause everything to crash.  The data will still need to be edited
to successfully import to Babase.